requirements.txt
click==8.1.7
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.27
pip==22.0.2
protobuf==5.27.1
PyJWT==2.8.0
setuptools==59.6.0
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.12.2
Werkzeug==2.0.2
